requirements.txt
adafruit-ampy==1.0.7
aioredis==1.3.1
alabaster==0.7.12
altgraph==0.17
appdirs==1.4.4
asgiref==3.2.7
astroid==2.4.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.3.0
autobahn==20.7.1
autolike==1.0.1
Automat==20.2.0
autopep8==1.5.2
Babel==2.8.0
boltons==19.3.0
bson==0.5.8
buildozer==1.2.0
cachier==1.4.2
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
daphne==2.5.0
dataclasses==0.6
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==2.2.14
django-allauth==0.42.0
django-channels==0.7.0
django-chartjs==2.2.1
django-cors-headers==3.3.0
django-developer-panel==0.1.2
django-environ==0.4.5
django-filter==2.3.0
django-grappelli==2.14.2
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-silk==4.0.1
django-widget-tweaks==1.4.8
djangorestframework==3.9.2
docopt==0.6.2
docutils==0.16
dpath==1.5.0
ecdsa==0.15
esptool==2.8
et-xmlfile==1.0.1
filelock==3.0.12
flake8==3.8.3
future==0.18.2
google-api-python-client==1.4.1
gprof2dot==2019.11.30
gunicorn==20.0.4
hiredis==1.1.0
httplib2==0.18.1
hyperlink==20.0.1
idna==2.9
imagesize==1.2.0
incremental==17.5.0
isort==4.3.21
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
micropy-cli==3.3.0
more-itertools==8.4.0
MouseInfo==0.1.3
msgpack==1.0.0
mysqlclient==1.4.6
numpy==1.18.4
oauth2client==4.1.3
oauthlib==3.1.0
openapi-codec==1.3.2
openpyxl==3.0.3
packaging==19.2
pandas==1.1.0
pathtools==0.1.2
pefile==2019.4.18
pexpect==4.8.0
Pillow==7.2.0
plotly==4.9.0
pluggy==0.13.1
portalocker==1.7.1
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.9.0
py2exe==0.9.2.2
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.50
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyGetWindow==0.0.8
Pygments==2.6.1
PyHamcrest==2.0.2
PyInstaller==3.6
pylint==2.5.3
pymongo==3.10.1
PyMsgBox==1.0.8
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.0
PyQt5==5.15.0
PyQt5-sip==12.8.0
PyRect==0.1.4
pyrsistent==0.16.0
PyScreeze==0.1.26
pyserial==3.4
pytest==5.4.3
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.14.0
python3-openid==3.2.0
PyTweening==1.0.3
pytz==2020.1
pywin32==228
pywin32-ctypes==0.2.0
QScintilla==2.11.5
questionary==1.5.2
requests==2.23.0
requests-oauthlib==1.3.0
requirements-parser==0.2.0
retrying==1.3.3
rsa==4.6
service-identity==18.1.0
sh==1.13.1
simplejson==3.17.0
six==1.14.0
snowballstemmer==2.0.0
Sphinx==3.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
toml==0.10.1
tqdm==4.48.0
Twisted @ file:///D:/Projects/core/Twisted-20.3.0-cp38-cp38-win_amd64.whl
txaio==20.4.1
uflash==1.2.4
Unipath==1.1
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.31
watchdog==0.10.3
wcwidth==0.2.5
whitenoise==5.1.0
wrapt==1.12.1
zope.interface==5.1.0
